2025/12/20 - 005

+++ blog/views.py
@@ -38,5 +38,9 @@ def article_edit(request, article_id):
 
 def article_delete(request, article_id):
     article = get_object_or_404(Article, id=article_id)
-    article.delete()
-    return redirect("blog:article_list")
+
+    if request.method == "POST":
+        article.delete()
+        return redirect("blog:article_list")
+
+    return render(request, "blog/article_delete.html", {"article": article})
